+++ .bitmap
@@ -525,6 +525,13 @@
         "mainFile": "index.ts",
         "rootDir": "scopes/toolbox/fs/hard-link-directory"
     },
+    "fs/linked-dependencies": {
+        "scope": "",
+        "version": "",
+        "defaultScope": "teambit.dependencies",
+        "mainFile": "index.ts",
+        "rootDir": "scopes/dependencies/fs/linked-dependencies"
+    },
     "generator": {
         "scope": "teambit.generator",
         "version": "0.0.1067",
@@ -656,13 +663,6 @@
         "version": "0.0.399",
         "mainFile": "index.ts",
         "rootDir": "scopes/component/legacy-component-log"
-    },
-    "linked-dependencies": {
-        "scope": "",
-        "version": "",
-        "defaultScope": "teambit.dependencies",
-        "mainFile": "index.ts",
-        "rootDir": "scopes/dependencies/linked-dependencies"
     },
     "linter": {
         "scope": "teambit.defender",
