+++ scopes/dependencies/fs/linked-dependencies/index.ts
