+++ scopes/dependencies/fs/linked-dependencies/linked-dependencies.ts
